+++ frontend/v1/domain/usecases/get-records-to-annotate-use-case.ts
@@ -25,11 +25,14 @@ export class GetRecordsToAnnotateUseCase {
     status: string,
     searchText: string
   ): Promise<void> {
-    const arrayOffset = page - 1;
+    const savedRecords = this.recordsStorage.get();
+
+    const { fromRecord, howMany } = savedRecords.getPageToFind(page, status);
 
     const getRecords = this.recordRepository.getRecords(
       datasetId,
-      arrayOffset,
+      fromRecord,
+      howMany,
       status,
       searchText
     );
@@ -95,7 +98,7 @@ export class GetRecordsToAnnotateUseCase {
           answer,
           suggestions,
           record.updated_at,
-          index + arrayOffset
+          index + page
         );
       }
     );