+++ src/argilla/__main__.py
@@ -18,6 +18,7 @@ import warnings
 from argilla.tasks import (
     database_app,
     datasets_app,
+    info_app,
     login_app,
     logout_app,
     server_app,
@@ -34,6 +35,7 @@ app = AsyncTyper(rich_help_panel=True, help="Argilla CLI", no_args_is_help=True)
 
 app.add_typer(database_app, name="database")
 app.add_typer(datasets_app, name="datasets")
+app.add_typer(info_app, name="info")
 app.add_typer(login_app, name="login")
 app.add_typer(logout_app, name="logout")
 app.add_typer(server_app, name="server")